Main file is added to project
